Moved files for user admin endpoint
