--- services/backend/src/core/admin/admin.js
@@ -1,1 +1,0 @@
-module.exports = require("./util/users");
